refactor(consolidate): extract chunk sizing + prompt to consolidate/chunking (D3 step 2)
Move CHARS_PER_TOKEN, PROMPT_OVERHEAD_TOKENS, DEFAULT_CONTEXT_LENGTH_TOKENS,
computeSafeChunkSize, and buildChunkPrompt into consolidate/chunking.ts (pure
token math + prompt-string assembly over MemoryEntry). clusterMemoriesBySimilarity
deliberately stays in the orchestrator (embedder/state-db coupling). Re-exported
from consolidate.ts; behavior-identical. 242 consolidate tests green.
